dev-requirements.txt
-r requirements.txt

django_debug_toolbar==1.3.0
ipython==3.1.0
pyinotify==0.9.5
